break out sequence persistence
